requirements.txt
requests
python-telegram-bot
Pillow
aiohttp
asyncpg